Make CoinSelections separate test suites not to mess with Jormungandr integration tests
