+++ lib/core-integration/src/Test/Integration/Scenario/API/Byron/Wallets.hs
@@ -16,10 +16,7 @@ module Test.Integration.Scenario.API.Byron.Wallets
 import Prelude
 
 import Cardano.Wallet.Api.Types
-    ( AddressAmount (..)
-    , ApiByronWallet
-    , ApiCoinSelectionOutput (..)
-    , ApiT (..)
+    ( ApiByronWallet
     , ApiUtxoStatistics
     , ApiWalletDiscovery (..)
     , DecodeAddress
@@ -28,29 +25,17 @@ import Cardano.Wallet.Api.Types
     , WalletStyle (..)
     )
 import Cardano.Wallet.Primitive.AddressDerivation
-    ( DerivationIndex (..)
-    , DerivationType (..)
-    , Index (..)
-    , PassphraseMaxLength (..)
-    , PassphraseMinLength (..)
-    , PaymentAddress
-    )
+    ( PassphraseMaxLength (..), PassphraseMinLength (..), PaymentAddress )
 import Cardano.Wallet.Primitive.AddressDerivation.Byron
     ( ByronKey )
 import Cardano.Wallet.Primitive.AddressDerivation.Icarus
     ( IcarusKey )
-import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
-    ( coinTypeAda, purposeBIP44 )
 import Cardano.Wallet.Primitive.SyncProgress
     ( SyncProgress (..) )
 import Control.Monad
     ( forM_, void )
 import Data.Generics.Internal.VL.Lens
-    ( view, (^.) )
-import Data.List
-    ( isPrefixOf )
-import Data.List.NonEmpty
-    ( NonEmpty ((:|)) )
+    ( (^.) )
 import Data.Maybe
     ( isJust, isNothing )
 import Data.Proxy
@@ -73,7 +58,6 @@ import Test.Integration.Framework.DSL
     , emptyIcarusWallet
     , emptyRandomWallet
     , emptyRandomWalletWithPasswd
-    , emptyWallet
     , eventually
     , expectErrorMessage
     , expectField
@@ -81,25 +65,19 @@ import Test.Integration.Framework.DSL
     , expectListSize
     , expectResponseCode
     , expectWalletUTxO
-    , fixtureIcarusWallet
-    , fixtureIcarusWalletWith
     , fixturePassphrase
     , fixturePassphraseEncrypted
     , genMnemonics
     , getFromResponse
     , json
-    , listAddresses
-    , minUTxOValue
     , request
     , rootPrvKeyFromMnemonics
-    , selectCoins
     , verify
     , walletId
     )
 import Test.Integration.Framework.TestData
     ( arabicWalletName
     , errMsg400NumberOfWords
-    , errMsg403NotAnIcarusWallet
     , errMsg403WrongPass
     , errMsg404NoWallet
     , kanjiWalletName
@@ -111,8 +89,6 @@ import Test.Integration.Framework.TestData
     )
 
 import qualified Cardano.Wallet.Api.Link as Link
-import qualified Data.List as L
-import qualified Data.List.NonEmpty as NE
 import qualified Data.Text as T
 import qualified Network.HTTP.Types.Status as HTTP
 
@@ -384,115 +360,6 @@ spec = describe "BYRON_WALLETS" $ do
             [ expectResponseCode @IO HTTP.status200
             , expectField (#name . #getApiT . #getWalletName) (`shouldBe` updatedName)
             ]
-
-    it "BYRON_COIN_SELECTION_00 - \
-        \No coin selection on Byron random" $ \ctx -> do
-        rnW <- emptyRandomWallet ctx
-        shW <- emptyWallet ctx
-        (addr:_) <- fmap (view #id) <$> listAddresses @n ctx shW
-        let payments = NE.fromList [ AddressAmount addr (Quantity minUTxOValue) ]
-        selectCoins @_ @'Byron ctx rnW payments >>= flip verify
-            [ expectResponseCode @IO HTTP.status403
-            , expectErrorMessage errMsg403NotAnIcarusWallet
-            ]
-
-    it "BYRON_COIN_SELECTION_01 - \
-        \A singleton payment is included in the coin selection output." $
-        \ctx -> do
-            source <- fixtureIcarusWallet ctx
-            target <- emptyWallet ctx
-            targetAddress : _ <- fmap (view #id) <$> listAddresses @n ctx target
-            let amt = Quantity minUTxOValue
-            let payment = AddressAmount targetAddress amt
-            let output = ApiCoinSelectionOutput targetAddress amt
-            let isValidDerivationPath path =
-                    ( length path == 5 )
-                    &&
-                    ( [ ApiT $ DerivationIndex $ getIndex purposeBIP44
-                      , ApiT $ DerivationIndex $ getIndex coinTypeAda
-                      , ApiT $ DerivationIndex $ getIndex @'Hardened minBound
-                      ] `isPrefixOf` NE.toList path
-                    )
-            selectCoins @_ @'Byron ctx source (payment :| []) >>= flip verify
-                [ expectResponseCode HTTP.status200
-                , expectField #inputs
-                    (`shouldSatisfy` (not . null))
-                , expectField #inputs
-                    (`shouldSatisfy` all
-                        (isValidDerivationPath . view #derivationPath))
-                , expectField #change
-                    (`shouldSatisfy` (not . null))
-                , expectField #change
-                    (`shouldSatisfy` all
-                        (isValidDerivationPath . view #derivationPath))
-                , expectField #outputs
-                    (`shouldBe` [output])
-                ]
-
-    it "BYRON_COIN_SELECTION_02 - \
-        \Multiple payments are all included in the coin selection output." $
-        \ctx -> do
-            let paymentCount = 10
-            source <- fixtureIcarusWallet ctx
-            target <- emptyWallet ctx
-            targetAddresses <- fmap (view #id) <$> listAddresses @n ctx target
-            let amounts = Quantity <$> [minUTxOValue ..]
-            let payments = NE.fromList
-                    $ take paymentCount
-                    $ zipWith AddressAmount targetAddresses amounts
-            let outputs =
-                    take paymentCount
-                    $ zipWith ApiCoinSelectionOutput targetAddresses amounts
-            selectCoins @_ @'Byron ctx source payments >>= flip verify
-                [ expectResponseCode HTTP.status200
-                , expectField #inputs (`shouldSatisfy` (not . null))
-                , expectField #change (`shouldSatisfy` (not . null))
-                , expectField
-                    #outputs (`shouldSatisfy` ((L.sort outputs ==) . L.sort))
-                ]
-
-    it "BYRON_COIN_SELECTION_03 - \
-        \Deleted wallet is not available for selection" $ \ctx -> do
-        icW <- emptyIcarusWallet ctx
-        shW <- emptyWallet ctx
-        (addr:_) <- fmap (view #id) <$> listAddresses @n ctx shW
-        let payments = NE.fromList [ AddressAmount addr (Quantity minUTxOValue) ]
-        _ <- request @ApiByronWallet ctx (Link.deleteWallet @'Byron icW) Default Empty
-        selectCoins @_ @'Byron ctx icW payments >>= flip verify
-            [ expectResponseCode @IO HTTP.status404
-            , expectErrorMessage (errMsg404NoWallet $ icW ^. walletId)
-            ]
-
-    it "BYRON_COIN_SELECTION_05 - \
-        \No change when payment fee eats leftovers due to minUTxOValue" $
-        \ctx -> do
-            source  <- fixtureIcarusWalletWith @n ctx [minUTxOValue, minUTxOValue]
-            target <- emptyWallet ctx
-
-            targetAddress:_ <- fmap (view #id) <$> listAddresses @n ctx target
-            let amt = Quantity minUTxOValue
-            let payment = AddressAmount targetAddress amt
-            let output = ApiCoinSelectionOutput targetAddress amt
-            let isValidDerivationPath path =
-                    ( length path == 5 )
-                    &&
-                    ( [ ApiT $ DerivationIndex $ getIndex purposeBIP44
-                      , ApiT $ DerivationIndex $ getIndex coinTypeAda
-                      , ApiT $ DerivationIndex $ getIndex @'Hardened minBound
-                      ] `isPrefixOf` NE.toList path
-                    )
-            selectCoins @_ @'Byron ctx source (payment :| []) >>= flip verify
-                [ expectResponseCode HTTP.status200
-                , expectField #inputs
-                    (`shouldSatisfy` (not . null))
-                , expectField #inputs
-                    (`shouldSatisfy` all
-                        (isValidDerivationPath . view #derivationPath))
-                , expectField #change
-                    (`shouldSatisfy` null)
-                , expectField #outputs
-                    (`shouldBe` [output])
-                ]
 
     it "BYRON_UTXO_01 - Wallet's inactivity is reflected in utxo" $ \ctx ->
         forM_ [ emptyRandomWallet, emptyIcarusWallet ] $ \emptyByronWallet -> do
